dressing-in-order/requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.1.1
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
autoviz==0.1.720
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
blis==0.7.10
bokeh==2.4.3
Bottleneck==1.3.7
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.0.1
comm==0.1.3
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
Cython==3.0.6
dask==2.30.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
emoji==2.5.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
entrypoints==0.4
exceptiongroup==1.1.3
executing==0.8.3
fastai==2.7.12
fastapi==0.103.1
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.17.1
fastprogress==1.0.3
ffmpy==0.3.1
filelock==3.12.4
Fiona==1.9.4.post1
Flask==3.0.0
Flask-Cors==3.0.10
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
gdown==4.7.1
gensim==4.3.2
geopandas==0.13.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googledrivedownloader==0.4
gradio==3.44.3
gradio_client==0.5.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
holoviews==1.14.9
htmlmin==0.1.12
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.1
hvplot==0.7.3
idna==3.4
ImageHash==4.3.1
imageio==2.33.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==7.7.5
isoduration==20.11.0
iterative-stratification==0.1.7
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.14
jsonpath-ng==1.5.3
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
jupyterlab_server==2.25.0
jupyterthemes==0.20.0
kaggle==1.5.13
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
lesscpy==0.15.1
levenshtein==0.12.0
libclang==16.0.6
locket==1.0.0
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
Metaphone==0.6
missingno==0.5.2
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.4
multimethod==1.9.1
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.6.3
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opendatasets==0.1.22
opt-einsum==3.3.0
orjson==3.9.7
outcome==1.2.0
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandas-dq==1.28
pandas-profiling==3.2.0
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
partd==1.4.0
pathy==0.10.2
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.1
plotly==5.15.0
ply==3.11
preshed==3.0.8
probableparsing==0.0.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.25.1
psutil==5.9.4
pure-eval==0.2.2
pyamg==5.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pyct==0.5.0
pydantic==1.10.9
pydub==0.25.1
pygame==2.5.2
Pygments==2.14.0
pyparsing==3.1.0
pyproj==3.6.0
pypyodbc==1.3.6
pyrsistent==0.19.3
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
python-stdnum==1.18
pytz==2023.3
pyviz-comms==2.3.2
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2020.11.13
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.5.1
starlette==0.27.0
statsmodels==0.14.0
sweetviz==2.1.4
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
text-unidecode==1.3
textblob==0.17.1
thinc==8.1.12
threadpoolctl==3.1.0
tifffile==2023.12.9
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
usaddress==0.5.10
uvicorn==0.23.2
varname==0.8.3
visions==0.7.4
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==3.6.4
wordcloud==1.9.2
wrapt==1.15.0
wsproto==1.2.0
xgboost==1.7.6
xlrd==2.0.1
yarl==1.9.2
